+++ source/extensions/lightspeed.trex.home.widget/lightspeed/trex/home/widget/tests/e2e/__init__.py
@@ -15,11 +15,6 @@
 * limitations under the License.
 """
 
-from .e2e.test_home_widget import TestHomeWidgetInvalidDepsFlow
-from .unit.test_home_widget import TestLoadWorkFile, TestRefreshRecentItemsDeferred
+from .test_home_widget import TestHomeWidgetInvalidDepsFlow
 
-__all__ = [
-    "TestHomeWidgetInvalidDepsFlow",
-    "TestLoadWorkFile",
-    "TestRefreshRecentItemsDeferred",
-]
+__all__ = ["TestHomeWidgetInvalidDepsFlow"]