+++ source/extensions/lightspeed.trex.utils.widget/docs/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.7.3]
+### Added
+- Added a shared invalid project `deps` rebuild dialog helper.
+
 ## [1.7.2]
 ### Added
 - Added a shared typed replacement asset file picker helper